UPDATED: stable version. The raspberry communication and the state machine work properly

+++ arduino/include/Tanque.h
@@ -22,8 +22,8 @@ private:
     uint8_t _pinTrigger;
 
 public:
-    uint8_t nivel = 0; // Nivel de fertilizante en el tanque (0–100, si luego lo calibras)
-    char tipo;         // Tipo de fertilizante por día
+    float nivel = 0; // Nivel de fertilizante en el tanque (0–100, si luego lo calibras)
+    char tipo;       // Tipo de fertilizante por día
 
     Tanque(uint8_t pinEcho,
            uint8_t pinTrigger,
